Fix issue with telegram user IDs longer than int32 due to recent telegram api change

+++ AdventureGameCore/InterfurCreations.AdventureGames.Telegram/DataObjects/GetUpdates.cs
@@ -8,6 +8,6 @@ namespace InterfurCreations.AdventureGames.Telegram.DataObjects
 {
     public class GetUpdates
     {
-        public int offset { get; set; }
+        public long offset { get; set; }
     }
 }
